==> scripts/04_velocity.Rmd <==
---
title: "Velocity Visualizing Script"
author: "Renjie Wu"
date: "`r Sys.Date()`"
output: 
  BiocStyle::html_document:
    toc: true
vignette: >
---

### YOU SHOULD HAVE RUN `scripts/03_model.sh` TO GET THE `velo.csv` FILE

### SET PREDICTED VELOCITY FILE & MAKE SURE the workspace is scripts/

```{r}
script_path <- dirname(rstudioapi::getActiveDocumentContext()$path)
setwd(script_path)

#TODO:
velofile <- "../test/model_test/normal_1/metrics/velo.csv"
```

### Load necessary libraries

```{r warning=FALSE,message=FALSE}
library(plotly)
library(RColorBrewer)

mytheme <- theme_bw() +
  theme(panel.grid.major = element_blank(),
        panel.grid.minor = element_blank(),
        panel.background = element_rect(colour = "black", size = 1),
        axis.ticks.length = unit(0.25, "cm"),
        axis.title = element_text(size = 14, face = "bold"),
        axis.text = element_text(size = 12),
        legend.title = element_text(size = 14),
        legend.text = element_text(size = 12),
        legend.position = "none")
```

### Read data

```{r}
logExpressionFiltered <- read.csv("../data/mESC/ExpressionData.csv",
                                  header = T, row.names = 1)
HayashiCellTypes <- read.csv("../data/mESC/celltype.csv",
                             header = T, row.names = 1)
```

### PCA

```{r}
# Compute PCA
pcaRes <- prcomp(t(logExpressionFiltered), scale. = FALSE)
rd1 <- as.data.frame(cbind(PC1 = pcaRes$x[,1], PC2 = pcaRes$x[,2], PC3 = pcaRes$x[,3]))

plot_ly(as.data.frame(pcaRes$x), x=~PC1, y= ~PC2,  color = HayashiCellTypes$Type, colors = brewer.pal(5,"Set1"))

write.csv(pcaRes$x[, 1:30], "../data/mESC/pca30.csv", quote = F)
```

### Real Velocity

```{r}
pt <- read.csv("../data/mESC/PseudoTime.csv", header = T, row.names = 1)
```

```{r}
k <- 0.2
weights <- rep(1/3, 3)
nxts <- c(32, 8, 2)
in_len <- 32
out_len <- 32

ord <- order(pt$PseudoTime)
orderedExpr <- logExpressionFiltered[, ord]

tmp <- rep(orderedExpr[, ncol(orderedExpr)], max(nxts)) %>% matrix(ncol = max(nxts)) %>% cbind(orderedExpr, .)

velo <- sapply(1:ncol(orderedExpr), function(i) {
  lapply(1:length(nxts), function(j) {
    (tmp[, i+nxts[j]] - tmp[, i]) * weights[j]
  }) %>% Reduce(`+`, .)
})

next_point <- velo * k + orderedExpr
proj_next <- predict(pcaRes, newdata = t(next_point))[, 1:2] %>% as.data.frame

df <- data.frame(cell_x = rd1$PC1[ord], cell_y = rd1$PC2[ord], next_x = proj_next$PC1, next_y = proj_next$PC2, c=HayashiCellTypes$Type[ord])

p1 <- ggplot(df[in_len:(nrow(df)-out_len), ], aes(x = cell_x, y = cell_y)) +
  geom_point(aes(color = c), size=2) +
  scale_color_manual(values = brewer.pal(5, "Set1")) +
  geom_segment(aes(xend=next_x, yend=next_y),
               arrow = arrow(length = unit(0.1, "inches"))) +
  mytheme +
  xlab("PC_1") + ylab("PC_2") +
  xlim(-32, 29) + ylim(-15, 16)
p1
ggsave("../fig/mESC/velo_true.png", width = 6, height = 4, dpi = 400)
```

### Predicted Velocity

```{r}
out_len <- 32
velo <- read.csv(velofile, header = F)
velo <- t(velo)

next_point <- velo * k + orderedExpr[, in_len:ncol(orderedExpr)]
proj_next <- predict(pcaRes, newdata = t(next_point))[, 1:2] %>% as.data.frame

df <- data.frame(cell_x = rd1$PC1[ord], cell_y = rd1$PC2[ord], c=HayashiCellTypes$Type[ord])[in_len:ncol(orderedExpr), ]
df$next_x <- proj_next$PC1
df$next_y <- proj_next$PC2

p2 <- ggplot(df[1:(nrow(df)-out_len), ], aes(x = cell_x, y = cell_y)) +
  geom_point(aes(color = c), size=2) +
  scale_color_manual(values = brewer.pal(5, "Set1")) +
  geom_segment(aes(xend=next_x, yend=next_y),
               arrow = arrow(length = unit(0.1, "inches"))) +
  mytheme +
  xlab("PC_1") + ylab("PC_2") +
  xlim(-32, 29) + ylim(-15, 16)
p2
ggsave("../fig/mESC/velo_pred.png", width = 6, height = 4, dpi = 400)
```

### NEXT, YOU MAY RUN `scripts/05_velo2pt.ipynb`
